labels tell you if they arent set

+++ src/Configurations/src/Configs/Configurable.java
@@ -9,8 +9,8 @@ public interface Configurable {
 
     @Retention(RetentionPolicy.RUNTIME)
     @interface Slider{
-        int min();
-        int max();
+        double min();
+        double max();
     }
 
     Configuration getConfiguration();
